jobs/api/requirements.txt
fastapi[all]==0.78.0
uvicorn[standard]==0.17.6
psycopg[binary,pool]==3.0.14
pytest==7.1.2
black==22.3.0
charset-normalizer==2.0.12
click==8.1.2
flake8==4.0.1
passlib[bcrypt]==1.7.4
pathspec==0.9.0
platformdirs==2.5.1
pycodestyle==2.8.0
pyflakes==2.4.0
python-multipart==0.0.5
python-dotenv==0.20.0
python-jose[cryptography]==3.2.0
requests==2.27.1
sqlparse==0.4.2
